requirements.txt
garth==0.4.47
python-dotenv==1.0.0
requests==2.32.3
cloudscraper==1.2.71
selenium==4.27.1
urllib3==2.2.3
python-json-logger==2.0.7
